requirements.txt
Flask
flask-mysqldb
python-dotenv
Flask-WTF
werkzeug